requirements.txt
aws-cdk-lib==2.215.0
constructs>=10.0.0,<11.0.0
boto3>=1.26.0
aws-lambda-powertools>=2.0.0
slack-sdk>=3.19.0